requirements.txt
Flask
pdfplumber
python-docx
pytesseract
pillow
scikit-learn